+++ src/components/NameListInput.vue
@@ -1,12 +1,12 @@
 <script setup lang="ts">
 import CardContainer from "./CardContainer.vue";
 
-const { modelValue } = defineProps<{ modelValue: string }>();
-const emit = defineEmits<{ (e: "update:modelValue", value: string): void }>();
+const modelValue = defineModel<string>();
+const id = crypto.randomUUID();
 </script>
 
 <template>
-  <label for="handles">
+  <label :for="id">
     <CardContainer bg-class="bg-indigo-200">
       <div class="col-span-full">
         <div class="block text-xl font-medium leading-6 text-gray-900">
@@ -14,18 +14,23 @@ const emit = defineEmits<{ (e: "update:modelValue", value: string): void }>();
         </div>
         <div class="mt-2">
           <textarea
-            :value="modelValue"
-            @input="
-              emit(
-                'update:modelValue',
-                (<HTMLTextAreaElement>$event.target).value,
-              )
-            "
-            id="handles"
+            v-model="modelValue"
+            :id="id"
             class="block min-h-48 w-full resize-none rounded-md border-0 p-2 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
           ></textarea>
         </div>
-        <p class="mt-3 text-sm leading-6 text-gray-600">One Handle per line.</p>
+        <div class="mt-2 flex">
+          <div class="text-sm leading-6 text-gray-600 flex-1 py-2">
+            One Handle per line.
+          </div>
+          <button
+            v-if="modelValue"
+            class="rounded-md font-semibold bg-red-400 p-2"
+            @click="modelValue = ''"
+          >
+            Clear
+          </button>
+        </div>
       </div>
     </CardContainer>
   </label>